+++ src/main/php/peer/http/HttpRequest.class.php
@@ -1,9 +1,8 @@
 <?php namespace peer\http;
 
 use peer\URL;
-use peer\Header;
 use peer\http\BasicAuthorization;
-use security\SecureString;
+use util\Secret;
 
 /**
  * Wrap HTTP/1.0 and HTTP/1.1 requests (used internally by the HttpConnection
@@ -39,7 +38,7 @@ class HttpRequest extends \lang\Object {
   public function setUrl(URL $url) {
     $this->url= $url;
     if ($url->getUser() && $url->getPassword()) {
-      $this->setHeader('Authorization', new BasicAuthorization($url->getUser(), new SecureString($url->getPassword())));
+      $this->setHeader('Authorization', new BasicAuthorization($url->getUser(), new Secret($url->getPassword())));
     }
     $port= $this->url->getPort(-1);
     $this->headers['Host']= [$this->url->getHost().(-1 == $port ? '' : ':'.$port)];